refactor: moving parsing of group authentication info out of index.js
